big structural fixes
