+++ frontend/src/App.css
@@ -29,3 +29,26 @@
     font-size: 60px;
     font-weight: bolder;
 }
+
+.PlayersBox {
+    grid-column-start: 1;
+    grid-column-end: 2;
+    grid-row-start: 2;
+    grid-row-end: 3;
+    border-style: solid;
+    border-color: var(--title-color);
+    border-radius: 10px;
+    padding: 15px;
+}
+
+.UserActionBox {
+    grid-column-start: 2;
+    grid-column-end: 3;
+    grid-row-start: 2;
+    grid-row-end: 3;
+    color: var(--text-color);
+    border-style: solid;
+    border-color: var(--title-color);
+    border-radius: 10px;
+    padding: 15px;
+}
