--- frontend/src/players/players.css
@@ -1,31 +1,0 @@
-.PlayersBox {
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 2;
-    grid-row-end: 3;
-    border-style: solid;
-    border-color: var(--title-color);
-    border-radius: 10px;
-    padding: 15px;
-}
-
-.PlayersTable {
-    width: 100%;
-    border: 1px solid;
-    border-collapse: collapse;
-    color: var(--text-color);
-    padding: 5px;
-}
-
-.PlayersTable td:last-child {
-    width: 80%;
-}
-
-.PlayersTable tr {
-    line-height: 10%;
-}
-
-.PlayersTable th {
-    line-height: 50%;
-    text-align: center;
-}
