+++ frontend/src/rounds/rounds.css
@@ -27,3 +27,24 @@ th, td {
     padding: 15px;
     border-color: black;
 }
+
+.PlayerPropositionsTable {
+    width: 100%;
+    border: 1px solid;
+    border-collapse: collapse;
+    color: var(--text-color);
+    padding: 5px;
+}
+
+.PlayerPropositionsTable td:last-child {
+    width: 80%;
+}
+
+.PlayerPropositionsTable tr {
+    line-height: 10%;
+}
+
+.PlayerPropositionsTable th {
+    line-height: 50%;
+    text-align: center;
+}
