+++ frontend/src/users/users.js
@@ -1,0 +1,32 @@
+import './users.css';
+
+
+let usersList = [
+    "Player1",
+    "Player2",
+    "Player3",
+    "Player4"
+]
+
+export function Users() {
+    return (
+        <div className="PlayersBox">
+            <table className="UsersTable">
+            <tr>
+                <th>
+                    Players
+                </th>
+            </tr>
+            {usersList.map((username, index) => {
+                return (
+                    <tr>
+                        <td>
+                            {username}
+                        </td>
+                    </tr>
+                )
+            })}
+            </table>
+        </div>
+    );
+}
